// src/packets/enums/TyreCompound.ts
export enum TyreCompound {
  /** Race 1 */
  TYRE_R1 = 0,

  /** Race 2 */
  TYRE_R2 = 1,

  /** Race 3 */
  TYRE_R3 = 2,

  /** Race 4 */
  TYRE_R4 = 3,

  /** Road super */
  TYRE_ROAD_SUPER = 4,

  /** Road normal */
  TYRE_ROAD_NORMAL = 5,

  /** Hybrid */
  TYRE_HYBRID = 6,

  /** Knobbly */
  TYRE_KNOBBLY = 7,

  TYRE_NUM = 8,

  /** Not changed */
  NOT_CHANGED = 255,
}